releasetools: ota_from_target_files: add recovery partition update
In standard Android systems, the recovery.img is a raw partition image
which is almost the same as boot.img except for the ramdisk that
changes. The update is then done as a patch applied at the startup right
after an update occurs. In our case, we chose to have the recovery files
separated and placedunder a FAT boot partition. So instead of of raw
patch, the recovery update is done during the rest of the update in two
steps:
1- copying same files as the boot partition
2- overwrite the ramdisk with the recovery console

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>